backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.7.0
channels==4.3.1
python-dotenv==1.1.1
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.10
daphne==4.2.1